Add testfx code for mainpage controller and scorecard controller
Add tests for both pages in the DiscoGolfAppTest. The code for the scorecard page will later be moved to
the other file
Add BeforeEach method for setting up the states in the scorecardController that are needed. This is done
using a Platform.runLater such that tesfx will run on correct thread
Comment out code for testing mainPageController temporarily because it is currently advantagious to only
test the scorecardController at this moment in the DiscoGolfAppTest file

Relates: #60 Authors: @jakobop and @ulrikyi