Add Leaderboard class
Need a Leaderboard class to hold and sort all the saved data
Create Leaderboard.java that holds alle the saved Scorecard
objects in a hashmap connected to their course, in sorted order.
Add getters to either return the whole hashmap or a leaderboard for
a specific couse name.
Add updateLeaderboard method that takes in a new Scorecard object and adds
it to relevant leaderboard, and sort the leaderboard.

Authors: @jakobop, @ulrikyi Relates: #82